+++ dictzip-lib/src/main/java/org/dict/zip/DictZipInputStream.java
@@ -82,6 +82,7 @@ public class DictZipInputStream extends InflaterInputStream {
     public DictZipInputStream(final RandomAccessInputStream in, final int size) throws IOException {
         super(in, new Inflater(true), size);
         header = readHeader();
+        readTrailer();
     }
 
     /**
@@ -263,36 +264,10 @@ public class DictZipInputStream extends InflaterInputStream {
     }
 
     /**
-     * Check gzip member trailer; CRC and length.
-     * @throws IOException when CRC error or total length error.
-     */
-    public void checkTrailer() throws IOException {
-        InputStream in = this.in;
-        int num = inf.getRemaining();
-        if (num > 0) {
-            in = new SequenceInputStream(
-                    new ByteArrayInputStream(buf, len - num, num), in);
-        }
-        long val = crc.getValue();
-        long crcValue = DictZipFileUtils.readUInt(in);
-        if (crcValue != val) {
-            throw new IOException(MessageFormat
-                    .format("Incorrect CRC: Computed CRC = %8x / From input %8x", val, crcValue));
-        }
-        long total = inf.getTotalOut();
-        long trailerTotal = DictZipFileUtils.readUInt(in);
-        if (trailerTotal != total) {
-            throw new IOException(MessageFormat
-                    .format("False number of uncompressed bytes: Computed size =%d / From input %d",
-                            total, trailerTotal));
-        }
-    }
-
-    /**
      * Reads GZIP member trailer.
      * @throws java.io.IOException If file I/O error
      */
-    public void readTrailer() throws IOException {
+    void readTrailer() throws IOException {
         if (in instanceof RandomAccessInputStream) {
             RandomAccessInputStream rain = (RandomAccessInputStream) in;
             compLength = rain.getLength();