+++ dictzip-lib/src/test/java/org/dict/zip/DictZipHeaderTest.java
@@ -202,6 +202,17 @@ public class DictZipHeaderTest {
     }
 
     /**
+     * Test of getMemberLength method.
+     */
+    @Test
+    public void testGetMemberLength() throws Exception {
+        System.out.println("getMemberLength");
+        String dataFile = this.getClass().getResource("/test.dict.dz").getFile();
+        DictZipHeader result = DictZipHeader.readHeader(dataFile);
+        assertEquals(result.getMemberLength(), 136856);
+    }
+
+    /**
      * Test of writeHeader method.
      *
      * @throws Exception if file I/O error occurred.